+++ Cross.toml
@@ -1,6 +1,9 @@
 [build]
 local-rust-root = false
 
+[build.env]
+passthrough = ["CURLOPT_PASSTHROUGH"]
+
 [target.aarch64-unknown-linux-gnu]
 mounts = [
     { source = "cross-cargo-cache", target = "/usr/local/cargo/registry", type = "volume" },
@@ -12,5 +15,6 @@ pre-build = [
     "apt-get update",
     "apt-get install --assume-yes libxkbcommon-dev:arm64 libwayland-dev:arm64 libx11-dev:arm64 libfontconfig1-dev:arm64 libudev-dev:arm64 libinput-dev:arm64 libgbm-dev:arm64"
 ]
-# Prevents skia-binding resume downloading feature
-env = { PASSTHROUGH = ["CURLOPT_PASSTHROUGH"], REWRITE = { "CURL_HOME" = "/tmp" } }
+
+[target.x86_64-unknown-linux-gnu.env]
+"CURL_HOME" = "/tmp"
