+++ kms_deploy.sh
@@ -7,11 +7,13 @@ TARGET_PATH="/home/admin"
 TARGET_TRIPLE="aarch64-unknown-linux-gnu"
 SOCKET="/tmp/ssh-socket-%r@%h:%p"
 
+FEATURE_FLAG="${RENDER_MODE:-linux-gpu}"
+
 RUN_IN_BACKGROUND=false
 
 echo "--- Building $APP_NAME ---"
 CROSS_CONTAINER_OPTS="--platform linux/amd64" cross build \
-    --features linux-gpu \
+    --features "$FEATURE_FLAG" \
     --package "$APP_NAME" \
     --target "$TARGET_TRIPLE" \
     --release
